Refactor tileindex into aoitools module
